day-01: solutions in rust
The part 2 was kind of a pain because some of the digit-names
overlapped and the built-in regexp parser does not seem to
support this, so I took a shortcut and encoded a few special
cases.

I need to start creating these solutions with #[cfg(test)] to
avoid the need to manually swap filenames in & out, but that's
how this one was done.