day-02: cubes of red, green, and blue
This was a lot easier than `day-01`, or at least part #2
didn't have much in the way of a twist.

I need to brush-up on more idiomatic use of hash-maps,
as the implementation here feels awkward.  There are
probably better ways of parsing the input than using
regexp as well.